channelをFFFF(65536)からFF(256)へ最大値を変更

+++ include/channel_map_simple_item.hpp
@@ -6,8 +6,8 @@
 namespace chmap {
     struct ChannelMapSimpleItem_FE {
         uint32_t id;
-        // id = (ip3rd << 16) | (ip4th << 8) | channel で初期化。それぞれ最大FF
-        ChannelMapSimpleItem_FE(uint8_t ip3rd, uint8_t ip4th, uint16_t ch) : id((uint32_t(ip3rd) << 16) | (uint32_t(ip4th) << 8) | uint32_t(ch) ) {}
+        // id = (ip3rd << 16) | (ip4th << 8) | channel で初期化。それぞれ最大FF、つまりid = 0x00FFFFFFまで
+        ChannelMapSimpleItem_FE(uint8_t ip3rd, uint8_t ip4th, uint8_t ch) : id((uint32_t(ip3rd) << 16) | (uint32_t(ip4th) << 8) | uint32_t(ch) ) {}
         public:
         void decode();
     };
